Change title name, npm i fontAwesome, edit styling for App and Navbar
