+++ public/index.html
@@ -24,6 +24,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
+    <link href="/your-path-to-fontawesome/css/fontawesome.css" rel="stylesheet">
     <title>Mode on Earth</title>
   </head>
   <body>
