+++ src/App.css
@@ -1,15 +1,26 @@
-/* .App {
-  text-align: center;
-} */
-
-.App__header {
-  background-color: #222;
-  background-image: url('https://unsplash.it/1000/300?image=823');
-  background-size: cover;
-  color: white;
-  padding: 10px 10px;
+.app {
+  height: 100%;
+  width: 100%;
+}
+.appHeader {  
+  display: flex; 
+  flex-direction: column;
+  align-items: center;
+}
+.appHeader h1 {
+  display: flex; 
+  justify-content: center;
+  /* width: 30%; */
+  margin: 0;
 }
 
+
+
+
+
+
+
+/* 
 .App__nav{
   width: 100%;
   list-style: none;
@@ -35,9 +46,9 @@
   background: none;
   cursor: pointer;
   color: white;
-}
+} */
 
-.button {
+/* .button {
   font-family: 'Poppins', sans-serif;
   text-transform: uppercase;
   background-color: #1A1B2A;
@@ -106,4 +117,5 @@
 
 .Flash__message--open {
   display: flex;
-}
+} */
+
