--- src/components/shopify/carts/Carts.css
@@ -1,84 +1,0 @@
-.Cart {
-  z-index: 9999;
-  position: fixed;
-  top: 0;
-  right: 0;
-  height: 100%;
-  width: 350px;
-  background-color: white;
-  display: flex;
-  flex-direction: column;
-  border-left: 1px solid #e5e5e5;
-  transform: translateX(100%);
-  transition: transform 0.15s ease-in-out;
-}
-
-@media (pointer:coarse) { // for all phones and tablets, take up full width (leave desktop with 350px)
-  .Cart {
-    width: 100%;
-  }
-}
-
-.Cart--open {
-  transform: translateX(0);
-}
-
-.Cart__close {
-  position: absolute;
-  right: 9px;
-  top: 8px;
-  font-size: 35px;
-  color: #999;
-  border: none;
-  background: transparent;
-  transition: transform 100ms ease;
-  cursor: pointer;
-}
-
-.Cart__header {
-  padding: 20px;
-  border-bottom: 1px solid #e5e5e5;
-  flex: 0 0 auto;
-  display: inline-block;
-}
-
-.Cart__line-items {
-  flex: 1 0 auto;
-  margin: 0;
-  padding: 20px;
-}
-
-.Cart__footer {
-  padding: 20px;
-  border-top: 1px solid #e5e5e5;
-  flex: 0 0 auto;
-}
-
-.Cart__checkout {
-  margin-top: 20px;
-  display: block;
-  width: 100%;
-}
-
-.Cart-info {
-  padding: 15px 20px 10px;
-}
-
-.Cart-info__total {
-  float: left;
-  text-transform: uppercase;
-}
-
-.Cart-info__small {
-  font-size: 11px;
-}
-
-.Cart-info__pricing {
-  float: right;
-}
-
-.pricing {
-  margin-left: 5px;
-  font-size: 16px;
-  color: black;
-}
