+++ src/components/shopify/lineItems/LineItems.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import './LineItems.css';
+// import './LineItems.css';
 
 class LineItems extends Component {
   constructor(props) {
