+++ src/styles/Shopify.css
@@ -1,4 +1,4 @@
-@import url('https://fonts.googleapis.com/css?family=Roboto:300,400,700');
+/* @import url('https://fonts.googleapis.com/css?family=Roboto:300,400,700');
 
 *, *:before, *:after {
   box-sizing: border-box;
@@ -11,7 +11,7 @@ html {
 body {
   margin: 0;
   padding: 0;
-  /* font-family: 'Roboto', sans-serif; */
+
   font-weight: 400;
 }
 
@@ -31,7 +31,7 @@ h2 {
   font-weight: 300;
   margin: 0;
   font-size: 2rem;
-}
+} */
 
 
 /* CUSTOMER AUTH
